230704
일반몬스터 패턴 1개
UI 수정 및 추가
Act / Mercy 선택에 따른 message box 출력

+++ 프레임워크 기본틀/230608/MonsterPhase.h
@@ -18,9 +18,9 @@ public:
 	virtual void Release(void) override;
 
 private:
-	DWORD		m_dwTime = GetTickCount64();
-	DWORD		m_preDeley = GetTickCount64();
-	DWORD		m_PostDelay = GetTickCount64();
+	DWORD		m_dwTime = (DWORD)GetTickCount64();
+	DWORD		m_preDeley = (DWORD)GetTickCount64();
+	DWORD		m_PostDelay = (DWORD)GetTickCount64();
 
 };
